rename obstacle layer
