+++ web/frontend/src/components/navigation/components/obstacle-layer.svelte
